requirements.txt
dj-database-url==2.1.0
django-admin-thumbnails==0.2.8
django-redis==5.3.0
djlint==1.31.0
gunicorn==21.2.0
pillow==9.5.0
pip-chill==1.0.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-decouple==3.8
qrcode==7.4.2
requests==2.31.0
